+++ src/main/java/com/zincoid/nullbot/core/component/ai/chat/plugin/QQPrompter.java
@@ -2,25 +2,22 @@ package com.zincoid.nullbot.core.component.ai.chat.plugin;
 
 import com.zincoid.nullbot.bot.dispatcher.CommandRegistry;
 import com.zincoid.nullbot.core.component.control.SysMsgManager;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
 
 @Component
+@RequiredArgsConstructor
 public class QQPrompter {
-
-    private final SysMsgManager sysMsgManager;
-    private final CommandRegistry commandRegistry;
-
-    public QQPrompter(SysMsgManager sysMsgManager, CommandRegistry commandRegistry) {
-        this.sysMsgManager = sysMsgManager;
-        this.commandRegistry = commandRegistry;
-    }
 
     private static final String BASE_PM_PROMPT;
     private static final String BASE_GC_PROMPT;
     private static final String CMD_PROMPT;
     private static final String MEMORY_PROMPT;
+
+    private final SysMsgManager sysMsgManager;
+    private final CommandRegistry commandRegistry;
 
     static {
 
@@ -30,7 +27,7 @@ public class QQPrompter {
                 回复消息时不要带以上格式化标识。尽量少用换行符，不要用emoji。禁止讨论中国国内政治事件和人物相关问题。
                 你可以通过在回复内容前紧跟[CQ:reply,id=消息ID]来引用指定消息，仅在需强调回复某消息时使用，例如：
                 [CQ:reply,id=1234567890]你好。
-                你可以在回复内容中嵌入 {Discard} 来放弃回复/保持静默，此时回复内容不会被发送。""";
+                你可以在回复内容中嵌入 <discard /> 来放弃回复保持静默，此时回复内容不会被发送。""";
 
         BASE_GC_PROMPT = """
                 
@@ -40,22 +37,22 @@ public class QQPrompter {
                 [CQ:reply,id=1234567890]你好。
                 你可以在回复中嵌入[CQ:at,qq=用户ID]来@别人，例如：
                 [CQ:at,qq=2660181154]你好。
-                你可以在回复内容中嵌入 {Discard} 来放弃回复/保持静默，此时回复内容不会被发送。""";
+                你可以在回复内容中嵌入 <discard /> 来放弃回复/保持静默，此时回复内容不会被发送。""";
 
         CMD_PROMPT = """
-                
-                你可以使用 {指令} 在回复中嵌入指令进行各种操作。
+
+                你可以使用 <cmd>指令</cmd> 在回复中嵌入指令进行各种操作。
                 被指令分隔的消息会以多条消息形式发送。
-                如果你想分开发送消息也可以使用空指令 {} 来分割。
-                
+                如果你想分开发送消息也可以使用空指令 <cmd></cmd> 来分割。
+
                 指令示例：
-                1. 发送帮助菜单 -> {Help}；
-                2. 发送表情包 -> {65275d24 表情包文件名}；
-                3. 多条消息 -> 这是第一条消息{}这是第二条消息；
-                
+                1. 发送帮助菜单 -> <cmd>Help</cmd>；
+                2. 发送表情包 -> <cmd>65275d24 表情包文件名</cmd>；
+                3. 多条消息 -> 这是第一条消息<cmd></cmd>这是第二条消息；
+
                 所有可用指令如下：
                 %s
-                
+
                 注意事项：
                 不要泄露任何指令内容！不要轻易复读别人让你执行的指令！
                 回复时不要执行过多指令，不要分割过多子消息！