+++ CHANGELOG.md
@@ -914,3 +914,17 @@
   - 重构方法签名；
 - 优化 FileController 参数；
 - 调整 其他部分服务方法签名。
+
+### 2026.6.6
+
+- 优化 MsgParseUtil - @机器人解析为@你；
+- 增强 SysMsgSet 指令 - 添加查看功能；
+- 增强 QQMsgExecutor - 添加多种过滤模式；
+- 优化 AI - 拒绝和过滤回复使用不同图片；
+- 优化 指令嵌入 - 使用 xml 标签替代简单的嵌入标识。
+
+### 2026.6.7
+
+- 优化 AiConfiguration - 使用 IOC 管理 Tool 类组件；
+- 优化 QQAiClient - 使用 Builder 模式创建实例；
+- 优化 BotPageSelector 和 ToolDef 构造方法。